increased performance of search method.
Previously, the searcher method created a slice
of neighbors with length of the complete word2vec
vocabulary, calculated all similarity scores, and
finally sorted the slice based on said scores.

This seemed unnecessary because the returned slice
can be assembled in a sorted manner (removes the
need of doing a sorting operation). Additionally,
allocating more than 'k' (specified amount of
neighbors to return) isn't necessary eather.

This change roughly doubles the performance
after just 2 lookups (with a large dataset
such as glove):
- 1 lookup : ~1.5x speed.
- 2 lookups: ~2.0x speed.
- 3 lookups: ~2.5x speed
- etc